ready to test runpod inference
